Fix so that Harvest handles non-200 responses
If Harvest made a request and the status of the response was not 200
(say, 404 or 500 or something), nothing would happen, and Harvest would
attempt to parse the response as if it were a 200. This is because it
looks for a "Status" header, which is a non-concept, when it just needs
to look for the response status.